move config.go to internal
